+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This page shows a detailed overview of the changes between versions without the 
 
 ## __WORK IN PROGRESS__
 
+- Enhancement: Show more details in the dashboard for software update states beside "Downloading"
 - Fix: Initialize "next node id" correctly when starting the server
 - Fix: Only execute custom attribute polling when the node is connected
 - Fix: Streamlines Hex-Number displaying in the Dashboard (Cluster-/Attribute-IDs)